multiplayer button + scrolling settings panel
