+++ _sass/minima/platformer-styles.scss
@@ -97,6 +97,8 @@ $tableBackground: #006bbd;
   width: 0px;
   padding-left: 0px; 
   padding-right: 0px; 
+  overflow-y: scroll; //allow scrolling
+  height: 75%;
 }
 
 .leaderboard {
