+++ assets/js/platformer2/SettingsControl.js
@@ -2,6 +2,7 @@
 import LocalStorage from "./LocalStorage.js";
 import GameEnv from "./GameEnv.js";
 import GameControl from "./GameControl.js";
+import Socket from "./Multiplayer.js";
 
 /* Coding Style Notes
  *
@@ -367,6 +368,21 @@ export class SettingsControl extends LocalStorage{
         return div;
     }
     
+    get multiplayerButton() {
+        const div = document.createElement("div");
+        div.innerHTML = "Multiplayer: "; // label
+    
+        const button = document.createElement("button"); // button for Multiplayer
+        button.innerText = String(Socket.shouldBeSynced);
+    
+        button.addEventListener("click", () => {
+            // dispatch event to update difficulty
+            button.innerText = String(Socket.changeStatus());
+        });
+    
+        div.append(button); // wrap button element in div
+        return div;
+    }
 
     /**
      * Static method to initialize the game settings controller and add the settings controls to the sidebar.
@@ -407,6 +423,10 @@ export class SettingsControl extends LocalStorage{
         // Get/Construct HTML input and event update for difficulty
         var difficultyInput = settingsControl.difficultyInput;
         document.getElementById("sidebar").append(difficultyInput);
+
+        // Get/Construct HTML button and event update for multiplayer
+        var multiplayerButton = settingsControl.multiplayerButton;
+        document.getElementById("sidebar").append(multiplayerButton);
 
 
         // Listener, isOpen, and function for sidebar open and close
